refactor(jellyfin): own audio/subtitle switching in the product service
switch_audio_track/switch_subtitle_track service commands absorb the
route handler cores (in-place mpv switch first, position-preserving
restart fallback). playback_service gains update_now_playing and a pause
reason passthrough so the switch paths write no transition globals, and
routes/jellyfin.py drops out of the NOW_PLAYING/SESSION_STATE writer pins.
Affected monkeypatches repointed to the service module.
