+++ docs/ARCHITECTURE_PHASE_4_JELLYFIN_INVENTORY.md
@@ -71,11 +71,10 @@ Service-level tests with fake receiver/player adapters are added in M7
 ## Jellyfin Function Listing
 
 <!-- BEGIN GENERATED JELLYFIN ROUTE LISTING (tests/test_jellyfin_inventory.py) -->
-### `routes/__init__.py` (16)
+### `routes/__init__.py` (15)
 
 - `_jellyfin_complete_ratio`
 - `_jellyfin_complete_remaining_sec`
-- `_jellyfin_emit_progress_hint`
 - `_jellyfin_emit_stopped_hint`
 - `_jellyfin_emit_stopped_payload`
 - `_jellyfin_integration_command_impl`
